+++ android/src/main/kotlin/com/kasem/receive_sharing_intent/ReceiveSharingIntentPlugin.kt
@@ -7,26 +7,29 @@ import android.media.MediaMetadataRetriever
 import android.media.ThumbnailUtils
 import android.net.Uri
 import android.provider.MediaStore
+import io.flutter.embedding.engine.plugins.FlutterPlugin
+import io.flutter.embedding.engine.plugins.activity.ActivityAware
+import io.flutter.embedding.engine.plugins.activity.ActivityPluginBinding
+import io.flutter.plugin.common.BinaryMessenger
 import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MethodCall
 import io.flutter.plugin.common.MethodChannel
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler
 import io.flutter.plugin.common.MethodChannel.Result
-import io.flutter.plugin.common.PluginRegistry
 import io.flutter.plugin.common.PluginRegistry.Registrar
+import io.flutter.plugin.common.PluginRegistry.NewIntentListener
 import org.json.JSONArray
 import org.json.JSONObject
 import java.io.File
 import java.io.FileOutputStream
 import java.net.URLConnection
-import android.os.Build
-import android.util.Log
 
+private const val MESSAGES_CHANNEL = "receive_sharing_intent/messages"
+private const val EVENTS_CHANNEL_MEDIA = "receive_sharing_intent/events-media"
+private const val EVENTS_CHANNEL_TEXT = "receive_sharing_intent/events-text"
 
-class ReceiveSharingIntentPlugin(val registrar: Registrar) :
-        MethodCallHandler,
-        EventChannel.StreamHandler,
-        PluginRegistry.NewIntentListener {
+class ReceiveSharingIntentPlugin : FlutterPlugin, ActivityAware, MethodCallHandler,
+        EventChannel.StreamHandler, NewIntentListener {
 
     private var initialMedia: JSONArray? = null
     private var latestMedia: JSONArray? = null
@@ -37,8 +40,26 @@ class ReceiveSharingIntentPlugin(val registrar: Registrar) :
     private var eventSinkMedia: EventChannel.EventSink? = null
     private var eventSinkText: EventChannel.EventSink? = null
 
-    init {
-        handleIntent(registrar.context(), registrar.activity().intent, true)
+    private var binding: ActivityPluginBinding? = null
+    private lateinit var applicationContext: Context
+
+    private fun setupCallbackChannels(binaryMessenger: BinaryMessenger) {
+        val mChannel = MethodChannel(binaryMessenger, MESSAGES_CHANNEL)
+        mChannel.setMethodCallHandler(this)
+
+        val eChannelMedia = EventChannel(binaryMessenger, EVENTS_CHANNEL_MEDIA)
+        eChannelMedia.setStreamHandler(this)
+
+        val eChannelText = EventChannel(binaryMessenger, EVENTS_CHANNEL_TEXT)
+        eChannelText.setStreamHandler(this)
+    }
+
+    override fun onAttachedToEngine(binding: FlutterPlugin.FlutterPluginBinding) {
+        applicationContext = binding.applicationContext
+        setupCallbackChannels(binding.binaryMessenger)
+    }
+
+    override fun onDetachedFromEngine(binding: FlutterPlugin.FlutterPluginBinding) {
     }
 
     override fun onListen(arguments: Any?, events: EventChannel.EventSink) {
@@ -55,44 +76,30 @@ class ReceiveSharingIntentPlugin(val registrar: Registrar) :
         }
     }
 
-    override fun onNewIntent(intent: Intent): Boolean {
-        handleIntent(registrar.context(), intent, false)
-        return false
-    }
-
+    // This static function is optional and equivalent to onAttachedToEngine. It supports the old
+    // pre-Flutter-1.12 Android projects. You are encouraged to continue supporting
+    // plugin registration via this function while apps migrate to use the new Android APIs
+    // post-flutter-1.12 via https://flutter.dev/go/android-project-migration.
+    //
+    // It is encouraged to share logic between onAttachedToEngine and registerWith to keep
+    // them functionally equivalent. Only one of onAttachedToEngine or registerWith will be called
+    // depending on the user's project. onAttachedToEngine or registerWith must both be defined
+    // in the same class.
     companion object {
-        private val MESSAGES_CHANNEL = "receive_sharing_intent/messages"
-        private val EVENTS_CHANNEL_MEDIA = "receive_sharing_intent/events-media"
-        private val EVENTS_CHANNEL_TEXT = "receive_sharing_intent/events-text"
-
         @JvmStatic
         fun registerWith(registrar: Registrar) {
-            // Detect if we've been launched in background
-            if (registrar.activity() == null) {
-                return
-            }
-
-            val instance = ReceiveSharingIntentPlugin(registrar)
-
-            val mChannel = MethodChannel(registrar.messenger(), MESSAGES_CHANNEL)
-            mChannel.setMethodCallHandler(instance)
-
-            val eChannelMedia = EventChannel(registrar.messenger(), EVENTS_CHANNEL_MEDIA)
-            eChannelMedia.setStreamHandler(instance)
-
-            val eChannelText = EventChannel(registrar.messenger(), EVENTS_CHANNEL_TEXT)
-            eChannelText.setStreamHandler(instance)
-
-            registrar.addNewIntentListener(instance)
+            val instance = ReceiveSharingIntentPlugin()
+            instance.applicationContext = registrar.context()
+            instance.setupCallbackChannels(registrar.messenger())
         }
     }
 
 
     override fun onMethodCall(call: MethodCall, result: Result) {
-        when {
-            call.method == "getInitialMedia" -> result.success(initialMedia?.toString())
-            call.method == "getInitialText" -> result.success(initialText)
-            call.method == "reset" -> {
+        when (call.method) {
+            "getInitialMedia" -> result.success(initialMedia?.toString())
+            "getInitialText" -> result.success(initialText)
+            "reset" -> {
                 initialMedia = null
                 latestMedia = null
                 initialText = null
@@ -103,7 +110,7 @@ class ReceiveSharingIntentPlugin(val registrar: Registrar) :
         }
     }
 
-    private fun handleIntent(context: Context, intent: Intent, initial: Boolean) {
+    private fun handleIntent(intent: Intent, initial: Boolean) {
         when {
             //Media is sent
             (intent.hasExtra(Intent.EXTRA_STREAM)) -> {
@@ -141,16 +148,16 @@ class ReceiveSharingIntentPlugin(val registrar: Registrar) :
         }
     }
 
-    private fun getMediaUris(context: Context, intent: Intent?): JSONArray? {
+    private fun getMediaUris(intent: Intent?): JSONArray? {
         if (intent == null) return null
 
-        return when {
-            intent.action == Intent.ACTION_SEND -> {
+        return when (intent.action) {
+            Intent.ACTION_SEND -> {
                 val uri = intent.getParcelableExtra<Uri>(Intent.EXTRA_STREAM)
-                val path = FileDirectory.getAbsolutePath(context, uri)
+                val path = uri?.let{ FileDirectory.getAbsolutePath(applicationContext, it) }
                 if (path != null) {
                     val type = getMediaType(path)
-                    val thumbnail = getThumbnail(context, path, type)
+                    val thumbnail = getThumbnail(path, type)
                     val duration = getDuration(path, type)
                     JSONArray().put(
                             JSONObject()
@@ -161,12 +168,13 @@ class ReceiveSharingIntentPlugin(val registrar: Registrar) :
                     )
                 } else null
             }
-            intent.action == Intent.ACTION_SEND_MULTIPLE -> {
+            Intent.ACTION_SEND_MULTIPLE -> {
                 val uris = intent.getParcelableArrayListExtra<Uri>(Intent.EXTRA_STREAM)
                 val value = uris?.mapNotNull { uri ->
-                    val path = FileDirectory.getAbsolutePath(context, uri) ?: return@mapNotNull null
+                    val path = FileDirectory.getAbsolutePath(applicationContext, uri)
+                            ?: return@mapNotNull null
                     val type = getMediaType(path)
-                    val thumbnail = getThumbnail(context, path, type)
+                    val thumbnail = getThumbnail(path, type)
                     val duration = getDuration(path, type)
                     return@mapNotNull JSONObject()
                             .put("path", path)
@@ -189,11 +197,11 @@ class ReceiveSharingIntentPlugin(val registrar: Registrar) :
         }
     }
 
-    private fun getThumbnail(context: Context, path: String, type: MediaType): String? {
+    private fun getThumbnail(path: String, type: MediaType): String? {
         if (type != MediaType.VIDEO) return null // get video thumbnail only
 
         val videoFile = File(path)
-        val targetFile = File(context.cacheDir, "${videoFile.name}.png")
+        val targetFile = File(applicationContext.cacheDir, "${videoFile.name}.png")
         val bitmap = ThumbnailUtils.createVideoThumbnail(path, MediaStore.Video.Thumbnails.MINI_KIND)
                 ?: return null
         FileOutputStream(targetFile).use { out ->
@@ -207,12 +215,36 @@ class ReceiveSharingIntentPlugin(val registrar: Registrar) :
         if (type != MediaType.VIDEO) return null // get duration for video only
         val retriever = MediaMetadataRetriever()
         retriever.setDataSource(path)
-        val duration = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION).toLongOrNull()
+        val duration = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_DURATION)?.toLongOrNull()
         retriever.release()
         return duration
     }
 
     enum class MediaType {
         IMAGE, VIDEO, FILE;
+    }
+
+    override fun onAttachedToActivity(binding: ActivityPluginBinding) {
+        this.binding = binding
+        binding.addOnNewIntentListener(this)
+        handleIntent(binding.activity.intent, true)
+    }
+
+    override fun onDetachedFromActivityForConfigChanges() {
+        binding?.removeOnNewIntentListener(this)
+    }
+
+    override fun onReattachedToActivityForConfigChanges(binding: ActivityPluginBinding) {
+        this.binding = binding
+        binding.addOnNewIntentListener(this)
+    }
+
+    override fun onDetachedFromActivity() {
+        binding?.removeOnNewIntentListener(this)
+    }
+
+    override fun onNewIntent(intent: Intent): Boolean {
+        handleIntent(intent, false)
+        return false
     }
 }