+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## 1.4.5
+
+* Update android to use compileSdkVersion 30
+
+## 1.4.4
+
+* Enabled null safety
+* Fix frozen app after sharing. Fix requires you to update your Sharing Extension Class (Check example project)
+* Fix regression regarding link propagation
+
+## 1.4.3
+
+* Fix suppressed link propagation to other modules on iOS
+
+## 1.4.2
+
+* Migrage to flutter android embedding v2
+* Fix crash while getting raw files path on Android
+
+## 1.4.1
+
+* Preserve original file name
+
+Note. Updating your Sharing Extension Class is required (please check that in the example project)
+
 ## 1.4.0
 
 * Added support for sharing any type of file